refactor(lib): extract to lib
